Version 1.7

+++ app/src/main/java/com/cs2340/smores/m4/model/PurityReportDBHandler.java
@@ -9,7 +9,7 @@ import android.database.sqlite.SQLiteOpenHelper;
 import java.util.ArrayList;
 
 /**
- * Created by smores on 2/22/17.
+ * Database handler to store PurityReport data between app sessions.
  */
 
 public class PurityReportDBHandler extends SQLiteOpenHelper {
@@ -22,12 +22,13 @@ public class PurityReportDBHandler extends SQLiteOpenHelper {
     private static final String KEY_NUMBER = "reportNumber";
     private static final String KEY_REAL_NAME = "realName";
     private static final String KEY_DATE = "dateCreated";
-    private static final String KEY_TYPE = "waterType";
     private static final String KEY_CONDITION = "condition";
     private static final String KEY_LONGITUDE = "longitude";
     private static final String KEY_LATITUDE = "latitude";
+    private static final String KEY_VIRUS_PPM = "virusPPM";
+    private static final String KEY_CONTAMINANT_PPM = "contaminantPPM";
 
-    public QualityReportDBHandler(Context context) {
+    public PurityReportDBHandler(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
     }
 
@@ -35,8 +36,9 @@ public class PurityReportDBHandler extends SQLiteOpenHelper {
     public void onCreate(SQLiteDatabase db) {
         String CREATE_USERS_TABLE = "CREATE TABLE " + TABLE_REPORTS + "("
                 + KEY_NUMBER + " INTEGER PRIMARY KEY," + KEY_REAL_NAME + " TEXT,"
-                + KEY_DATE + " TEXT," + KEY_TYPE + " TEXT," + KEY_CONDITION + " TEXT,"
-                + KEY_LONGITUDE + " DOUBLE," + KEY_LATITUDE + " DOUBLE" + ")";
+                + KEY_DATE + " TEXT," + KEY_CONDITION + " TEXT,"
+                + KEY_LONGITUDE + " DOUBLE," + KEY_LATITUDE + " DOUBLE,"
+                + KEY_VIRUS_PPM + " INTEGER," + KEY_CONTAMINANT_PPM + " INTEGER" + ")";
         db.execSQL(CREATE_USERS_TABLE);
     }
 
@@ -47,19 +49,22 @@ public class PurityReportDBHandler extends SQLiteOpenHelper {
     }
 
     void addReport(PurityReport report) {
-        SQLiteDatabase db = this.getWritableDatabase();
+        if (report != null) {
+            SQLiteDatabase db = this.getWritableDatabase();
 
-        ContentValues values = new ContentValues();
-        values.put(KEY_NUMBER, report.getReportNumber());
-        values.put(KEY_REAL_NAME, report.getName());
-        values.put(KEY_DATE, report.getTimeCreated());
-        values.put(KEY_TYPE, report.getWaterType());
-        values.put(KEY_CONDITION, report.getCondition());
-        values.put(KEY_LONGITUDE, report.getLongitude());
-        values.put(KEY_LATITUDE, report.getLatitude());
+            ContentValues values = new ContentValues();
+            values.put(KEY_NUMBER, report.getReportNumber());
+            values.put(KEY_REAL_NAME, report.getName());
+            values.put(KEY_DATE, report.getTimeCreated());
+            values.put(KEY_CONDITION, report.getCondition());
+            values.put(KEY_LONGITUDE, report.getLongitude());
+            values.put(KEY_LATITUDE, report.getLatitude());
+            values.put(KEY_VIRUS_PPM, report.getVirusPPM());
+            values.put(KEY_CONTAMINANT_PPM, report.getContaminantPPM());
 
-        db.insert(TABLE_REPORTS, null, values);
-        db.close();
+            db.insert(TABLE_REPORTS, null, values);
+            db.close();
+        }
     }
 
     public ArrayList<PurityReport> getPurityReports() {
@@ -97,9 +102,11 @@ public class PurityReportDBHandler extends SQLiteOpenHelper {
     }
 
     void removeReport(PurityReport report) {
-        SQLiteDatabase db = this.getWritableDatabase();
-        db.delete(TABLE_REPORTS, KEY_NUMBER + " = ?",
-                new String[] { String.valueOf(report.getReportNumber()) });
-        db.close();
+        if (report != null) {
+            SQLiteDatabase db = this.getWritableDatabase();
+            db.delete(TABLE_REPORTS, KEY_NUMBER + " = ?",
+                    new String[]{String.valueOf(report.getReportNumber())});
+            db.close();
+        }
     }
 }
