+++ app/src/main/java/com/cs2340/smores/m4/model/UserDBHandler.java
@@ -24,6 +24,8 @@ public class UserDBHandler extends SQLiteOpenHelper {
     private static final String KEY_PHONE_NUMBER = "phoneNumber";
     private static final String KEY_ADDRESS = "address";
     private static final String KEY_USER_TYPE = "userType";
+    private static final String KEY_IS_BANNED = "isBanned";
+    private static final String KEY_IS_LOCKED = "isLocked";
 
     public UserDBHandler(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
@@ -33,9 +35,9 @@ public class UserDBHandler extends SQLiteOpenHelper {
     public void onCreate(SQLiteDatabase db) {
         String CREATE_USERS_TABLE = "CREATE TABLE " + TABLE_USERS + "("
                 + KEY_USERNAME + " TEXT PRIMARY KEY," + KEY_REAL_NAME + " TEXT,"
-                + KEY_PASSWORD + " TEXT," + KEY_USER_TYPE + " TEXT,"
-                + KEY_EMAIL + " TEXT," + KEY_PHONE_NUMBER +
-                " TEXT," + KEY_ADDRESS + " TEXT" + ")";
+                + KEY_PASSWORD + " TEXT," + KEY_EMAIL + " TEXT," + KEY_PHONE_NUMBER
+                + " TEXT," + KEY_ADDRESS + " TEXT," + KEY_USER_TYPE + " INTEGER,"
+                + KEY_IS_BANNED + " BOOLEAN," + KEY_IS_LOCKED + " BOOLEAN" + ")";
         db.execSQL(CREATE_USERS_TABLE);
     }
 
@@ -55,7 +57,9 @@ public class UserDBHandler extends SQLiteOpenHelper {
         values.put(KEY_EMAIL, user.getEmail());
         values.put(KEY_PHONE_NUMBER, user.getPhoneNumber());
         values.put(KEY_ADDRESS, user.getAddress());
-        values.put(KEY_USER_TYPE, user.type());
+        values.put(KEY_USER_TYPE, user.getType());
+        values.put(KEY_IS_BANNED, user.isBanned());
+        values.put(KEY_IS_LOCKED, user.isLocked());
 
         db.insert(TABLE_USERS, null, values);
         db.close();
@@ -70,33 +74,21 @@ public class UserDBHandler extends SQLiteOpenHelper {
 
         if (cursor.moveToFirst()) {
             User user;
-            String name;
-            String username;
-            String password;
-            String email;
-            String phoneNumber;
-            String address;
+            String name, username, password, email, phoneNumber, address;
+            boolean isBanned, isLocked;
+            int userType;
             do {
-                name = cursor.getString(1);
                 username = cursor.getString(0);
+                name = cursor.getString(1);
                 password = cursor.getString(2);
-                email = cursor.getString(4);
-                phoneNumber = cursor.getString(5);
-                address = cursor.getString(6);
-                switch (cursor.getString(3)) {
-                    case "Admin":
-                        user = new Admin(name, username, password, email, phoneNumber, address);
-                        break;
-                    case "Manager":
-                        user = new Manager(name, username, password, email, phoneNumber, address);
-                        break;
-                    case "Worker":
-                        user = new Worker(name, username, password, email, phoneNumber, address);
-                        break;
-                    default:
-                        user = new Customer(name, username, password, email, phoneNumber, address);
-                        break;
-                }
+                email = cursor.getString(3);
+                phoneNumber = cursor.getString(4);
+                address = cursor.getString(5);
+                userType = Integer.parseInt(cursor.getString(6));
+                isBanned = Boolean.parseBoolean(cursor.getString(7));
+                isLocked = Boolean.parseBoolean(cursor.getString(8));
+                user = new User(name, username, password, email,
+                        phoneNumber, address, isBanned, isLocked, userType);
                 users.add(user);
             } while (cursor.moveToNext());
         }
@@ -105,10 +97,32 @@ public class UserDBHandler extends SQLiteOpenHelper {
         return users;
     }
 
+    public void updateUser(User user, String oldUsername) {
+        if ((user != null) && (oldUsername != null)) {
+            SQLiteDatabase db = this.getWritableDatabase();
+            db.delete(TABLE_USERS, KEY_USERNAME + " = ?",
+                    new String[]{user.getUsername()});
+            ContentValues values = new ContentValues();
+            values.put(KEY_USERNAME, user.getUsername());
+            values.put(KEY_REAL_NAME, user.getRealName());
+            values.put(KEY_PASSWORD, user.getPassword());
+            values.put(KEY_EMAIL, user.getEmail());
+            values.put(KEY_PHONE_NUMBER, user.getPhoneNumber());
+            values.put(KEY_ADDRESS, user.getAddress());
+            values.put(KEY_USER_TYPE, user.getType());
+            values.put(KEY_IS_BANNED, user.isBanned());
+            values.put(KEY_IS_LOCKED, user.isLocked());
+            db.insert(TABLE_USERS, null, values);
+            db.close();
+        }
+    }
+
     void removeUser(User user) {
-        SQLiteDatabase db = this.getWritableDatabase();
-        db.delete(TABLE_USERS, KEY_USERNAME + " = ?",
-                new String[] { String.valueOf(user.getUsername()) });
-        db.close();
+        if (user != null) {
+            SQLiteDatabase db = this.getWritableDatabase();
+            db.delete(TABLE_USERS, KEY_USERNAME + " = ?",
+                    new String[]{user.getUsername()});
+            db.close();
+        }
     }
 }
